+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## Platform Version 0.9.25 - UNRELEASED
+* Set timestamp in Records while producing. ([#2288](https://github.com/infinyon/fluvio/issues/2288))
 
 
 ## Platform Version 0.9.24 - 2022-04-21